Remove unessecary type precision

+++ src/Device/Resistance/Resistance.h
@@ -16,7 +16,7 @@ namespace devuino
             /* Make a reading of a analog input from (T signal) by powering on a resistor from the (T power) pin.
              * That to prevent for example powerdraw and corrosion on water sensors. */
 
-            Resistance(const T signal, const T power, const bool debounce = false, const uint8_t iterations = 10, const uint8_t bitresolution = 10)
+            Resistance(const T signal, const T power, const bool debounce = false, const unsigned int iterations = 10, const uint8_t bitresolution = 10)
                 : signal(signal), power(power), iterations(iterations), InputAnalog(bitresolution, debounce)
             {
                 this->signal.initiate(pin::Mode::InputAnalog);
@@ -49,7 +49,7 @@ namespace devuino
           protected:
             T signal;
             T power;
-            const uint8_t iterations;
+            const unsigned int iterations;
         };
     }
 }
